Safari-611.1.10/LayoutTests/ietestcenter/Javascript/11.13.1-1-3-expected.txt

11.13.1-1-3 (TestCases/chapter11/11.13/11.13.1/11.13.1-1-3.js) simple assignment throws ReferenceError if LeftHandSide is not a reference (boolean)

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS ES5Harness.preconditionPassed is true
FAIL ES5Harness.testPassed should be true (of type boolean). Was undefined (of type undefined).
PASS successfullyParsed is true

TEST COMPLETE
